requirements.txt
fastapi==0.111
pandas==2.2.2
numpy==1.26.4
langchain==0.1.16
llama-cpp-python==0.2.74
faiss-cpu==1.8.0
torch==2.3.0
sentence-transformers==2.7.0
langchain-community==0.0.33
rank_bm25==0.2.2
loguru==0.7.2
wget==3.2
scikit-learn==1.4.2
